requirements.txt
streamlit
langchain
langchain-groq
langchain-openai
python-dotenv
requests

